Fixed some mistakes at LexAnalyzer, added some stuff.
Fixed a mistake in the transitions matrix that made string literals
containing escape sequences rejected by the automaton. Fixed a
similar mistake with characters, when receiving a '\', it would go
to the wrong state (10 instead of 11).

Added const char* getTypeString() method to the Token class. This
method returns a string representing the token type.

Also fixed the fact that the compiler didn't call the method readFile
at all.